changement base de données

+++ controllers/socket.js
@@ -3,8 +3,8 @@ const usbSerial = '/dev/ttyUSB0';
 const SerialPort = require('serialport');
 const jwt = require('jsonwebtoken');
 
-const User = require('../Models/User');
-const Stack = require('../Models/Stack')
+const User = require('../Models/user');
+const Stack = require('../Models/stack')
 
 const keyToken = 'P7H}9C7ccv^Sk7Yia0C1Te1o3g2gqTt6EmuyIi.g8(}iQLM+sGX5577&0SF)e50)kjDomBt6Ns^MAHZ7#3Tq{87~2m=UInz7L05@XwC2dJHS5FAX:P?3@*:2ALII4G@Hf!Uc1akX?:xMm6bt<(b27VW80lcVf&;d99CVfNS+0Ni28Q{q8!7Y5}(C48zO@x5C8-PHn/j=Bc00998C{VK:cE09GS5_B10R8YR3?077r~v89hQI6p{Kydu65|0$py&c{Pdl[70FL|B%);uib4*dQ5@6!^%6^$j1vhn2%5H=E02!6224[nFiF5,&ctI-~s(7@L&:,~0e281ki>1A7FS7:7$2KTfe3u787a^8-qH4Yu6R96a@)p*25811~|RG,9UpsA$;1hW7[(/OZb5)6rN~:swMTam7/h!{^PjWE0<2WK$+$i?}p:%e;3g~A%:q)zZs$lL9$A>Z>qF}[4wUYf#0&*Mq8csI$?5F2mG@o^ZhsMa]wRDSqY#m0[j@lt/$zoW7';
 let token = "";
@@ -16,9 +16,10 @@ let rankController = -1;
 
 exports.connected = (ws) => {
     ws.on('connection', async () => {
-
+        console.log("connected")
     })
     ws.on('message', async (msg) => {
+        console.log(msg)
         let data = JSON.parse(msg)
 
         //connection init
@@ -64,8 +65,7 @@ exports.connected = (ws) => {
                 mega.on('data', function (data) {
                     console.log('data received: ' + data);
                 });
-            }
-            else {
+            } else {
                 //update rankController
                 rankController = await getRankController();
                 ws.send(JSON.stringify({
@@ -93,14 +93,14 @@ exports.connected = (ws) => {
                                 let userRank = count.length === 0 ? 0 : count[0].rank + 1;
 
                                 let stack = new Stack({
-                                    idUser: user._id,
+                                    idUser: user.id,
                                     rank: userRank % 5000
                                 });
                                 stack.save()
                                     .then(async () => {
                                         ws.send(JSON.stringify({
                                             token: jwt.sign({
-                                                idUser: user._id,
+                                                idUser: user.id,
                                                 rank: userRank,
                                             }, keyToken)
                                         }));
@@ -135,17 +135,20 @@ exports.connected = (ws) => {
         }
     });
     ws.on('close', () => {
+        console.log("close")
         if (mega) {
             mega.write("01002011001200");
             mega.close();
             mega = null;
         }
-        const decodedToken = jwt.verify(token, keyToken);
-        Stack.deleteOne({idUser: decodedToken.idUser})
-            .then(() => {
-                User.deleteOne({_id: decodedToken.idUser})
-                    .then(() => console.log("close"))
-            })
+        if (token) {
+            const decodedToken = jwt.verify(token, keyToken);
+            Stack.deleteOne({idUser: decodedToken.idUser})
+                .then(() => {
+                    User.deleteOne({id: decodedToken.idUser})
+                        .then(() => console.log("close"))
+                })
+        }
     })
 }
